New vertical block funcionality.
From Eric Welser:

I've included a patch for MP 5.0.2 to add vertical cut/copy/paste. This
allow the user to select text in rectangular blocks, and use the standard
editing functions on that block. This is patterned on similar functionality
in SlickEdit, Vim and many other editors.

Press CTRL-B to enter vertical selection mode, and move the cursor to mark
the selection rectngle. The existing edit functions (copy, cut, paste) will
behave as expected with the highlighted text. Text added to the clipboard
from external applications are used with normal selection mode.